new version of repo server: swagger for controller, node 14, consistent require instead of import, all server stuff in /app folder
