+++ Dockerfile.dev
@@ -5,9 +5,13 @@ ENV HTTP_PORT=7060
 ENV ADMIN_PORT=7061
 ENV WEBHOOK_PORT=7062
 ENV EXTERNAL_HOST="3.141.51.96"
+
+#dev wallet
 ENV WALLET_KEY="Repo.Agent"
 ENV WALLET_NAME="Repo.Agent"
+
 ENV GENESIS_FILE="/home/indy/resources/genesis.txt"
+ENV TAILS_URL="http://vswrepo.com:6543"
 
 EXPOSE $PORT
 
@@ -15,18 +19,12 @@ USER root
 RUN curl -sL https://deb.nodesource.com/setup_14.x | bash - && \
   apt-get install nodejs -y
 
-COPY package.json ./
-COPY package-lock.json ./
+COPY ./app/package.json ./
+COPY ./app/package-lock.json ./
 COPY ./app/ ./
-COPY ./app/resources/.indy_dev ./.indy_client
+COPY ./app/resources/.indy_client ./.indy_client
 
 RUN touch logs/agent.logs
 RUN npm install && npm cache clean --force
 
-# HEALTHCHECK --interval=1m --timeout=3s \
-#  CMD curl \
-#      --fail \
-#      http://localhost:$PORT/adminRoutes/health || \
-#      exit 1
-
-CMD node app.js
+CMD node --trace-warnings ./bin/www
