--- app/routes/adminRoutes.js
@@ -1,41 +1,0 @@
-
-import express from 'express'
-import { Router } from 'express'
-import axios from "axios";
-
-const DEFAULT_INTERNAL_HOST = `${process.env.DOCKERHOST}` || hostname.docker.internal;
-const DEFAULT_EXTERNAL_HOST = `${process.env.EXTERNAL_HOST}` || `${process.env.DOCKERHOST}`;
-
-const HTTP_PORT = `${process.env.HTTP_PORT}` || 8060;
-const ADMIN_PORT = `${process.env.ADMIN_PORT}` || 8061;
-const WEBHOOK_PORT = `${process.env.WEBHOOK_PORT}` || 8062;
-
-const ADMIN_URL = `http://${DEFAULT_EXTERNAL_HOST}:${ADMIN_PORT}`
-
-// health
-
-let adminRoutes = Router()
-  .get("/health", async (req, response) => {
-    var data = " ";
-    var config = {
-      method: "get",
-      url: `${ADMIN_URL}/status`,
-      headers: { "Content-Type": "application/json" },
-      data: data,
-    };
-    axios(config)
-      .then(function (res) {
-        console.log("/adminRoutes/health");
-        //console.log(res)
-        var results = res.data;
-        console.log(results);
-        response.json(results);
-      })
-      .catch((error) => {
-        console.error(error.response);
-        response.status(500).send(error).end();
-      });
-  });
-
-export default  adminRoutes ;
-
