+++ app/routes/users.js
@@ -1,9 +1,9 @@
 var express = require('express');
 var router = express.Router();
 
-/* GET home page. */
+/* GET users listing. */
 router.get('/', function(req, res, next) {
-  res.render('index', { title: 'VSW REPO' });
+  res.send('respond with a resource');
 });
 
 module.exports = router;
