--- app/test/common.test.js
@@ -1,3 +1,0 @@
-
-
-
