--- app/variables.js
@@ -1,17 +1,0 @@
-
-// TODO
-let variables;
-
-function getRandomInt(max) {
-    return Math.floor(Math.random() * Math.floor(max));
-}
-  
-let seed = "my_seed_00000000000000000000" + "8876";
-let did = "";
-
-variables = {
-    seed : seed,
-    did: did
-};
-
-export default variables;
