--- package.json
@@ -1,30 +1,0 @@
-{
-  "name": "vsw-repo",
-  "version": "0.1.0",
-  "description": "Verifiable Software repository controller",
-  "main": "index.js",
-  "type": "module",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/verifiablesoftware/vsw.git",
-    "directory": "vsw-repo"
-  },
-  "author": "Futurewei Technologies",
-  "license": "Apache-2.0",
-  "bugs": {
-    "url": "https://github.com/verifiablesoftware/vsw/issues"
-  },
-  "homepage": "https://github.com/verifiablesoftware/vsw#readme",
-  "dependencies": {
-    "aws-sdk": "^2.795.0",
-    "axios": "^0.21.1",
-    "body-parser": "^1.19.0",
-    "express": "^4.17.1",
-    "json": "^10.0.0",
-    "morgan": "^1.10.0",
-    "pg": "^8.5.1"
-  }
-}
