+++ repo-provision-aws-dev.sh
@@ -3,6 +3,6 @@ aca-py provision \
 --endpoint http://3.141.51.96:8070 \
 --genesis-file "/home/jte/git/vsw-repo/app/resources/genesis.txt" \
 --wallet-type indy \
---wallet-name Repo.Agent \
---wallet-key Repo.Agent \
+--wallet-name Repo.Dev \
+--wallet-key Repo.Dev \
 --seed my_seed_000000000000000000007777
